packages/trustix-nix-reprod/pyproject.toml
[tool.poetry]
name = "trustix-nix-reprod"
version = "0.1.0"
description = ""
authors = ["adisbladis <adam.hose@tweag.io>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
ijson = "^3.1.3"
python-dateutil = "^2.8.1"
grpcio = "^1.34.0"
pynix = {path = "../pynix"}
trustix-python = {path = "../trustix-python"}
tortoise-orm = {extras = ["asyncpg", "accel"], version = "^0.16.18"}
aerich = "^0.5.0"
aiofiles = "^0.6.0"
async_lru = "^1.0.2"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
fastapi = {extras = ["standard"], version = "^0.63.0"}
Jinja2 = "^2.11.3"
aiohttp = {extras = ["speedups"], version = "^3.7.4"}
python-multipart = "^0.0.5"
python-Levenshtein = "^0.12.2"
requests = "^2.25.1"
orjson = "^3.5.1"
aiocache = {extras = ["redis", "memcached"], version = "^0.11.1"}

[tool.poetry.dev-dependencies]
mypy = "^0.812"
black = "^20.8b1"
flake8 = "^3.8.4"
pytest = "^6.1.2"
ipython = "^7.19.0"
grpcio-tools = "^1.34.0"
watchdog = "^2.0.2"
ansicolors = "^1.1.8"

[tool.poetry.scripts]
trustix_nix_reprod = 'trustix_nix_reprod.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | node_modules
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
